+++ app/matrix/page.tsx
@@ -1,0 +1,5 @@
+import MatrixResult from '../pages/MatrixResult';
+
+export default function MatrixPage() {
+  return <MatrixResult />;
+}